fix: restructure components folder
